requirements.txt
# Bibliotecas principais para análise de dados
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Visualização de dados
matplotlib==3.7.2
seaborn==0.12.2

# MLOps - Rastreamento de experimentos
mlflow==2.5.0

# API REST
fastapi==0.101.0
uvicorn==0.23.2
pydantic==2.1.1

# Autenticação JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate limiting (throttling)
slowapi==0.1.8

# Interface web
streamlit==1.25.0

# Requisições HTTP
requests==2.31.0

# Variáveis de ambiente
python-dotenv==1.0.0

# Testes
pytest==7.4.0

# Utilitários
joblib==1.3.1
PyYAML==6.0.1